api-challenge/requirements.txt
aniso8601==8.1.0
astroid==2.4.2
atomicwrites==1.4.0
attrs==20.3.0
click==7.1.2
colorama==0.4.4
Flask==1.1.2
Flask-RESTful==0.3.8
importlib-metadata==3.1.1
iniconfig==1.1.1
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.19.3
packaging==20.7
pandas==1.1.5
pluggy==0.13.1
py==1.9.0
pylint==2.6.0
pyparsing==2.4.7
pytest==6.1.2
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
toml==0.10.2
typed-ast==1.4.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
